Fix some rebasing fallouts
