+++ graphdatascience/graph/graph_export_runner.py
@@ -2,7 +2,6 @@ from typing import Any, Dict
 
 from pandas import Series
 
-from ..caller_base import CallerBase
 from ..error.illegal_attr_checker import IllegalAttrChecker
 from ..error.uncallable_namespace import UncallableNamespace
 from .graph_object import Graph
@@ -28,7 +27,7 @@ class GraphExportCsvRunner(IllegalAttrChecker):
         return self._export_call(G, config)
 
 
-class GraphExportCsvEndpoints(CallerBase, UncallableNamespace, IllegalAttrChecker):
+class GraphExportCsvEndpoints(UncallableNamespace, IllegalAttrChecker):
     @property
     def csv(self) -> GraphExportCsvRunner:
         self._namespace += ".csv"
@@ -36,7 +35,7 @@ class GraphExportCsvEndpoints(CallerBase, UncallableNamespace, IllegalAttrChecke
         return GraphExportCsvRunner(self._query_runner, self._namespace, self._server_version)
 
 
-class GraphExportRunner(CallerBase, IllegalAttrChecker):
+class GraphExportRunner(IllegalAttrChecker):
     def __call__(self, G: Graph, **config: Any) -> "Series[Any]":
         return self._export_call(G, config)
