+++ graphdatascience/ignored_server_endpoints.py
@@ -56,4 +56,5 @@ IGNORED_SERVER_ENDPOINTS = {
     "gds.util.isFinite",
     "gds.util.isInfinite",
     "gds.alpha.create.cypherdb",
+    "gds.alpha.graph.project",  # TODO: Figure out how to support this well
 }
