+++ graphdatascience/topological_lp/topological_lp_alpha_runner.py
@@ -4,7 +4,7 @@ from ..error.illegal_attr_checker import IllegalAttrChecker
 from ..error.uncallable_namespace import UncallableNamespace
 
 
-class TopologicalLPAlphaRunner(CallerBase, UncallableNamespace, IllegalAttrChecker):
+class TopologicalLPAlphaRunner(UncallableNamespace, IllegalAttrChecker):
     def _run_standard_function(self, node1: int, node2: int, config: Dict[str, Any]) -> float:
         query = f"""
         MATCH (n1) WHERE id(n1) = {node1}
